Move OSGi aware functionallity in BindingManagement to Activator.
BindingManagement is NOT aware of the existance of bundles.
Instead, it holds a set with capability classes known to it, and another one with aplications.
It uses these sets for the binding mechanism.

These sets are updated whenever a capabilitiesAdded/Removed or applicationsAdded/Removed callbacks are received.

In an OSGi environment, available classes may change when bundles enter or leave the system.
Thus, the Activator (an OSGi aware class) is now:
- tracking bundles
- detecting changes in the availability of capabilities and applications
- and notifying BindingManagement for these changes

This separation allows the creation of unitary tests for BindingManagement.