Prepare BindingManagement to use the resource tree.
Implement IBindingManagement, IBindingManagementEventListener used to manage the resource and capabilities tree.

The logic has been implemented but it is not yet being called when resources, capabilityclasses and applicationclasses are added/removed in the system.

TODO: Replace all model with the new one (now data is duplicated). Remove unused methods.
TODO: Modify ResourceMonitoringFilter to report which capability instance is launching the event.