ApplicationInstances init/cancel services when resolved/unresolved.
BindingManagement triggers service initialization when the ApplicationInstance is resoldev.
It triggers service cancelation when ApplicationInstance gets unresolved.

CapabilityInstances follow the same behaviour.
Services in CapabilityInstances are no longer initialized when the instance is bound, but when it gets resolved.
Services in CapabilityInstances are no longer cancelled when the instance is unbound, but when it gets unresolved.