Modify equals to fix resource lookup.
Resources are meant to be equals only when they represent the same resource.
Thus, the UnitResource id must be included in the equals

Units represent the same concept when they have same name and same type.
Thus Units are equal when they contain UnitResources with same name and type (but ignoring their resource id.)