Virtual resources created by SlicingCapability are bound to virtual network IRootResourceProvider.
When processing a VIR, the NetworkManagement manually binds sliced resources to the newly created virtual network.
In order to achieve that, createSlice must not bind the newly created slice to itseld (thus, @Adds Resource annotation is not desired.), otherwise
the manually bind attempt will silently fail.

In order for the manual bind to succeed, addsResource() should be called with
the capability instance instead of the proxy.
This is achieved by using getCapabilityInstance service from serviceProvider.