Avoid concurrent modification exception in iterator for knownCapabilities.
This patch avoids concurrent modification exception in resourceAdded when knownCapabilities changes while iterating over it.
It solves the issue by forcing knownCapabilities not changing meanwhile.
This is achieved by means of a mutex.

This patch uses a mutex to "synchronize" uses of knownCapabilities inside BindingManagement.
This way, we can avoid the potential race condition between a thread adding resources
and another adding or removing capabilities in the system.